refactor: use instance pattern in FileSystem and Inode; refactor: replace Inode as a trait; refactor: add redoxfs entry for Inode trait;
Redoxfs need a git deps(rv64) for redox_syscall, not available yet

+++ os/src/syscall/gui.rs
@@ -26,8 +26,8 @@ pub fn sys_framebuffer() -> isize {
     let mut inner = current_process.inner_exclusive_access();
     inner.memory_set.push(
         MapArea::new(
-            (FB_VADDR as usize).into(),
-            (FB_VADDR + len as usize).into(),
+            FB_VADDR.into(),
+            (FB_VADDR + len).into(),
             MapType::Linear(pn_offset),
             MapPermission::R | MapPermission::W | MapPermission::U,
         ),
